Update doctest expected output for numpy 2.0's changed scalar repr (e.g. np.int64(1) instead of 1)

+++ deepchem/data/data_loader.py
@@ -472,7 +472,7 @@ class UserCSVLoader(CSVLoader):
     >>> len(dataset)
     2
     >>> dataset.X[0, 0]
-    1
+    np.int64(1)
 
     The difference between `UserCSVLoader` and `CSVLoader` is that our
     descriptors (our features) have already been computed for us, but are spread
