+++ deepchem/feat/vocabulary_builders/grover_vocab.py
@@ -529,7 +529,7 @@ class GroverAtomVocabTokenizer(Featurizer):
     >>> atom_tokenizer = GroverAtomVocabTokenizer(file.name)
     >>> mol = Chem.MolFromSmiles('CC(=O)C')
     >>> atom_tokenizer.featurize([(mol, mol.GetAtomWithIdx(0))])[0]
-    2
+    np.int64(2)
 
     Parameters
     ----------
@@ -563,7 +563,7 @@ class GroverBondVocabTokenizer(Featurizer):
     >>> bond_tokenizer = GroverBondVocabTokenizer(file.name)
     >>> mol = Chem.MolFromSmiles('CC(=O)C')
     >>> bond_tokenizer.featurize([(mol, mol.GetBondWithIdx(0))])[0]
-    2
+    np.int64(2)
 
     Parameters
     ----------
